[fix] BroadcastChannel 및 relogin layout에서 실행되도록 수정
